Use img tag to lazy load speaker images
Speaker pictures use an `<img>` tag with lazy loading enabled to
minimize data fetching on page load. As the image was no longer part of
the background, the styling needed updates. As a bonus flourish we now
apply filters to the image rather than swapping it out for a solid
color.